move stuff into framework/effects
