+++ lib/framework.rb
@@ -1,3 +1,4 @@
+require "framework/effects"
 require "framework/abilities"
 require "framework/cards"
