+++ lib/framework/abilities.rb
@@ -1,5 +1,4 @@
 require "framework/abilities/definition"
-require "framework/abilities/dsl"
 
 module Framework
   module Abilities
