+++ lib/framework/effects.rb
@@ -1,0 +1,6 @@
+require "framework/effects/definition"
+
+module Framework
+  module Effects
+  end
+end
